rest_app/requirements.txt
Django==1.9
django-filter==0.11.0
django-rest-auth==0.6.0
djangorestframework==3.3.2
Markdown==2.6.5
psycopg2==2.6.1
six==1.10.0
wheel==0.24.0
